🗺️ Atlas: Split Lexicon and Conversion blobs
Extracted `src/morphology/lexicon.rs` and `src/semantic/conversion.rs` into module directories to resolve The Blob anti-pattern. Moved `LEXICON` static data into `lexicon/data.rs` and large inline test blocks into dedicated `tests.rs` files.
